replace more bf16 with half

+++ ggml/src/ggml-cuda/convert.cu
@@ -785,7 +785,7 @@ to_fp16_nc_cuda_t ggml_get_to_fp16_nc_cuda(ggml_type type) {
 to_bf16_nc_cuda_t ggml_get_to_bf16_nc_cuda(ggml_type type) {
     switch (type) {
         case GGML_TYPE_F32:
-            return convert_unary_cuda<float, nv_bfloat16>;
+            return convert_unary_cuda<float, half>;
         case GGML_TYPE_Q4_0:
             return dequantize_block_cuda<QK4_0, QR4_0, dequantize_q4_0>;
         case GGML_TYPE_Q4_1:
@@ -797,7 +797,7 @@ to_bf16_nc_cuda_t ggml_get_to_bf16_nc_cuda(ggml_type type) {
         case GGML_TYPE_Q8_0:
             return dequantize_block_cuda<QK8_0, QR8_0, dequantize_q8_0>;
         case GGML_TYPE_F16:
-            return convert_unary_cuda<half, nv_bfloat16>;
+            return convert_unary_cuda<half, half>;
         default:
             return nullptr;
     }
@@ -818,7 +818,7 @@ to_fp32_nc_cuda_t ggml_get_to_fp32_nc_cuda(ggml_type type) {
         case GGML_TYPE_Q8_0:
             return dequantize_block_cuda<QK8_0, QR8_0, dequantize_q8_0>;
         case GGML_TYPE_BF16:
-            return convert_unary_cuda<nv_bfloat16, float>;
+            return convert_unary_cuda<half, float>;
         default:
             return nullptr;
     }
